Move the IMU processing functions into seperate files
